requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4

argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
blis==0.7.11

boto3==1.34.147
botocore==1.34.147

cachetools==5.4.0
catalogue==2.0.10
cer==1.2.0


charcut==1.1.1
chardet==5.2.0
charset-normalizer
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comet-ml==3.47.0
comm==0.2.2



confection==0.1.5
configobj==5.0.9
contourpy==1.2.1

cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
daisy==1.1.1
dataclasses==0.6
Dbias==0.1.5
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1

dnspython==2.6.1
docker==7.1.0
dulwich==0.22.1
email_validator==2.2.0
everett==3.1.0
executing==2.0.1
fairlearn==0.10.0
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
funlib.geometry==0.3.0
funlib.math==0.1
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.3

iniconfig==2.0.0
ipapi==1.0.4
ipykernel==6.29.5
ipython==8.10.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jiwer==3.0.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25

jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
Levenshtein==0.25.1
libclang==18.1.1

lightgbm==4.4.0
lightning-utilities==0.11.6
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mauve-text==0.4.0
mdurl==0.1.2

mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.42.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0

pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs 
plotly==5.23.0
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pureml-evaluate==0.1.8
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0

pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyngrok==7.2.0
pyparsing==3.1.2
PyPDF2==3.0.1

pytest==8.3.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.5
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
rsa==4.9

s3transfer==0.10.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentry-sdk==2.16.0
setuptools==71.1.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
terminado==0.18.1
thinc==8.2.5
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.8
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
transforms==0.2.1

typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.10
ultralytics-thop==2.0.9
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wurlitzer==3.1.1
xlrd==2.0.1

